test function bodies in parser
A naive test of a function with a body failed because we were
not actually looking for or accepting semicolons, but were generating
them on printing a thing back out. Changed to require semicolons
between statements. This logic may be wrong; for instance, if
you have a trailing semicolon on a body, the String() method won't
yield the original text exactly.